+++ tests/Integration/Commands/Install-SqlDscPowerBIReportServer.Integration.Tests.ps1
@@ -29,12 +29,9 @@ BeforeAll {
     Import-Module -Name $script:moduleName -Force -ErrorAction 'Stop'
 }
 
-Describe 'Repair-SqlDscBIReportServer' -Tag @('Integration_PowerBI') {
+Describe 'Install-SqlDscPowerBIReportServer' -Tag @('Integration_PowerBI') {
     BeforeAll {
         Write-Verbose -Message ('Running integration test as user ''{0}''.' -f $env:UserName) -Verbose
-
-        # Starting the Power BI Report Server service prior to running tests.
-        Start-Service -Name 'PowerBIReportServer' -Verbose -ErrorAction 'Stop'
 
         $script:temporaryFolder = Get-TemporaryFolder
 
@@ -42,33 +39,39 @@ Describe 'Repair-SqlDscBIReportServer' -Tag @('Integration_PowerBI') {
         $powerBIReportServerExecutable = Join-Path -Path $script:temporaryFolder -ChildPath 'PowerBIReportServer.exe'
     }
 
-    It 'Should have the BI Report Server service running' {
-        $getServiceResult = Get-Service -Name 'PowerBIReportServer' -ErrorAction 'Stop'
-
-        $getServiceResult.Status | Should -Be 'Running'
-    }
-
-    Context 'When repairing BI Report Server' {
+    Context 'When installing Power BI Report Server' {
+        # cSpell: ignore PBIRS
         It 'Should run the command without throwing' {
-            # Set splatting parameters for Repair-SqlDscBIReportServer
-            $repairSqlDscBIReportServerParameters = @{
+            # Set splatting parameters for Install-SqlDscPowerBIReportServer
+            $installSqlDscBIReportServerParameters = @{
                 AcceptLicensingTerms = $true
                 MediaPath            = $powerBIReportServerExecutable
-                LogPath              = Join-Path -Path $script:temporaryFolder -ChildPath 'SSRS_Repair.log'
+                InstallFolder        = 'C:\Program Files\PBIRS'
+                Edition              = 'Developer'
+                LogPath              = Join-Path -Path $script:temporaryFolder -ChildPath 'PowerBIReportServer_Install.log'
                 SuppressRestart      = $true
                 Verbose              = $true
-                ErrorAction          = 'Stop'
                 Force                = $true
             }
 
-            $null = Repair-SqlDscBIReportServer @repairSqlDscBIReportServerParameters
+            $null = Install-SqlDscPowerBIReportServer @installSqlDscBIReportServerParameters -ErrorAction 'Stop'
         }
 
-        It 'Should still have a Power BI Report Server service running after repair' {
-            $getServiceResult = Get-Service -Name 'PowerBIReportServer' -ErrorAction 'Stop'
+        It 'Should have installed Power BI Report Server' {
+            # Validate the Power BI Report Server installation
+            $reportServerService = Get-Service -Name 'PowerBIReportServer'
 
-            $getServiceResult | Should -Not -BeNullOrEmpty
-            $getServiceResult.Status | Should -Be 'Running'
+            $reportServerService | Should -Not -BeNullOrEmpty
+            $reportServerService.Status | Should -Be 'Running'
+        }
+
+        It 'Should stop the Power BI Report Server service' {
+            # Stop the Power BI Report Server service to save memory on the build worker
+            $stopServiceResult = Stop-Service -Name 'PowerBIReportServer' -Force -PassThru -Verbose -ErrorAction 'Stop'
+
+            write-verbose -Message ($stopServiceResult | Out-String) -Verbose
+
+            $stopServiceResult.Status | Should -Be 'Stopped'
         }
     }
 }